requirements.txt
alabaster==0.7.11
alembic==1.7.4
altgraph==0.17.2
anaconda-client==1.7.2
anaconda-navigator==2.1.0
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
blinker==1.4
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.7
conda==4.10.3
conda-build==3.15.1
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262352237/work
conda-repo-cli @ file:///tmp/build/80754af9/conda-repo-cli_1620168426516/work
conda-token @ file:///tmp/build/80754af9/conda-token_1620076980546/work
constantly==15.1.0
contextlib2==0.5.5
cplex==20.1.0.1
cryptography @ file:///C:/ci/cryptography_1615505408273/work
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
datashape==0.5.4
DBUtils==2.0.2
decorator==4.3.0
defusedxml==0.5.0
distributed==1.23.1
docutils==0.14
dominate==2.6.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.8
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
glob2==0.6
graphviz==0.16
greenlet==1.1.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==4.7.1
importlib-resources==5.4.0
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython_genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==2.0.1
jdcal==1.4
jedi==0.12.1
Jinja2==3.0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==2.2.3
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
mysqlclient==2.0.3
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.15.1
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pefile==2021.9.3
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==5.4.7
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.4
pycurl==7.44.0
pydotplus==2.0.2
pyflakes==2.0.0
Pygments==2.2.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT @ file:///C:/ci/pyjwt_1619651810943/work
pylint==2.1.1
PyMySQL==1.0.2
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==5.3.1
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
requests==2.19.1
rope==0.11.0
ruamel_yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.4.22
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==18.7.0
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.23
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.4.1
win-unicode-console==0.5
win_inet_pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
WTForms==2.3.3
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
xmltodict @ file:///Users/ktietz/demo/mc3/conda-bld/xmltodict_1629301980723/work
zict==0.1.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
